apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
sqlalchemy==2.0.29
alembic==1.13.1
anthropic==0.25.0
python-multipart==0.0.9
slowapi==0.1.9
PyJWT==2.8.0
python-docx==1.1.0
weasyprint==61.2
celery==5.3.6
redis==5.0.4
email-validator==2.1.1
